+++ apps/admin/src/app/modules/admin-categories/admin-categories.module.ts
@@ -1,11 +1,15 @@
-import { Module } from '@nestjs/common'
+import {forwardRef, Module} from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import {AdminCategory} from "./admin-categories.entity";
 import {AdminCategoriesController} from "./admin-categories.controller";
 import {AdminCategoriesService} from "./admin-categories.service";
+import {AdminMediaModule} from "../admin-media/admin-media.module";
 
 @Module({
-  imports: [TypeOrmModule.forFeature([AdminCategory])],
+  imports: [
+    TypeOrmModule.forFeature([AdminCategory]),
+    forwardRef(() => AdminMediaModule)
+  ],
   controllers: [AdminCategoriesController],
   providers: [AdminCategoriesService],
   exports: [AdminCategoriesService]