+++ libs/interfaces/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/pagination-object.interface'
 export * from './lib/builder-type.enum'
+export * from './lib/token-payload.interface'